pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "unifi-protect-alerts"
version = "0.1.0"
description = "UniFi Protect → Webhook → FastAPI → OpenAI Vision → Pushover"
authors = [{name="Your Name"}]
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.8",
]

[project.optional-dependencies]
dev = ["ruff>=0.5"]
